Moving each type definition to a separate file.
