+++ src/System.Devices.Gpio/System.Devices.Gpio/PinMode.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.Devices.Gpio
+{
+    public enum PinMode
+    {
+        Input,
+        Output,
+        InputPullDown,
+        InputPullUp
+    }
+}
