+++ src/System.Devices.Gpio/System.Devices.Gpio/PinNumberingScheme.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.Devices.Gpio
+{
+    public enum PinNumberingScheme
+    {
+        Board,
+        BCM
+    }
+}
