+++ src/System.Devices.Gpio/System.Devices.Gpio/PinValue.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.Devices.Gpio
+{
+    public enum PinValue
+    {
+        Low = 0,
+        High = 1
+    }
+}
